+++ lib/core/request.js
@@ -217,7 +217,10 @@ class Request {
     this.completed = true
     if (channels.trailers.hasSubscribers) {
       channels.trailers.publish({
-        request: { ...this, bodyReceived: this[kHandler].bodyReceived },
+        request: this,
+        response: {
+          body: this[kHandler].bodyReceived
+        },
         trailers
       })
     }
