+++ test/diagnostics-channel/get.js
@@ -114,9 +114,8 @@ diagnosticsChannel.channel('undici:request:headers').subscribe(({ request, respo
 
 let endEmitted = false
 diagnosticsChannel.channel('undici:request:trailers').subscribe(({ request, trailers }) => {
-  delete request.bodyReceived
   t.equal(request.completed, true)
-  t.same(_req, request)
+  t.equal(_req, request)
   // This event is emitted after the last chunk has been added to the body stream,
   // not when it was consumed by the application
   t.equal(endEmitted, false)
